Update 10.4.1 Writing a recursive function for factorial First write the base case, then add the recursive case.c
Creating a recursive function
Creating a recursive function can be accomplished in two steps.

Write the base case -- Every recursive function must have a case that returns a value without performing a recursive call. That case is called the base case. A programmer may write that part of the function first, and then test. There may be multiple base cases.
Write the recursive case -- The programmer then adds the recursive case to the function.
The following illustrates a simple function that computes the factorial of N (i.e. N!). The base case is N = 1 or 1! which evaluates to 1. The base case is written as  if (N <= 1) { fact = 1; }. The recursive case is used for N > 1, and written as else { fact = N * NFact( N - 1 ); }.

PARTICIPATION ACTIVITY
10.4.1: Writing a recursive function for factorial: First write the base case, then add the recursive case.


1

2

int NFact(int N) {
   int factResult;

   if (N <= 1) { // Base case
      factResult = 1;
   }
   else {        // Recursive case
      factResult = N * NFact(N - 1);
   }
   return factResult;
}

// main(): Get N, print NFact(N)
int NFact(int N) {
   int factResult;

   if (N <= 1) { // Base case
      factResult = 1;
   }
   // FIXME: Finish
   return factResult;
}

// main(): Get N, print NFact(N)
Enter N: 1
N! is: 1

Enter N: 6
N! is: 720
Enter N: 1
N! is: 1
int NFact(int N) {

   int factResult;
   if (N <= 1) { // Base case
      factResult = 1;
   }
   // FIXME: Finish
   return factResult;
}

// main(): Get N, print NFact(N)
Next the recursive case, which calls itself, is written and tested.
If N is greater than 1, then the NFact() function returns N * NFact(N - 1).
Captions

The base case, which returns a value without performing a recursive call, is written and tested first. If N is less than or equal to 1, then the NFact() function returns 1.
Next the recursive case, which calls itself, is written and tested. If N is greater than 1, then the NFact() function returns N * NFact(N - 1).

Feedback?
A common error is to not cover all possible base cases in a recursive function. Another common error is to write a recursive function that doesn't always reach a base case. Both errors may lead to infinite recursion, causing the program to fail.

Typically, programmers will use two functions for recursion. An "outer" function is intended to be called from other parts of the program, like the function int CalcFactorial(int inVal). An "inner" function is intended only to be called from that outer function, for example a function int CalcFactorialHelper(int inVal). The outer function may check for a valid input value, e.g., ensuring inVal is not negative, and then calling the inner function. Commonly, the inner function has parameters that are mainly of use as part of the recursion, and need not be part of the outer function, thus keeping the outer function more intuitive.