Create 10.5.1: Fibonacci sequence step-by-step.c
10.5 Recursive math functions
Fibonacci sequence
Recursive functions can solve certain math problems, such as computing the Fibonacci sequence. The Fibonacci sequence is 0, 1, 1, 2, 3, 5, 8, 13, 21, 34, etc.; starting with 0, 1, the pattern is to compute the next number by adding the previous two numbers.

Below is a program that outputs the Fibonacci sequence values step-by-step, for a user-entered number of steps. The base case is that the program has output the requested number of steps. The recursive case is that the program needs to compute the number in the Fibonacci sequence.

OUTPUT:
This program outputs the
Fibonacci sequence step-by-step,
starting after the first 0 and 1.

How many steps would you like? 10
0
1
0+1= 1
1+1= 2
1+2= 3
2+3= 5
3+5= 8
5+8= 13
8+13= 21
13+21= 34
21+34= 55
34+55= 89