Create 10.8 C example: Recursively output permutations.c
10.8.1: Recursively output permutations.
The below program prints all permutations of an input string of letters, one permutation per line. Ex: The six permutations of "cab" are:

cab
cba
acb
abc
bca
bac
Below, the PermuteString function works recursively by starting with the first character and permuting the remainder of the string. The function then moves to the second character and permutes the string consisting of the first character and the third through the end of the string, and so on.

Run the program and input the string "cab" (without quotes) to see that the above output is produced.
Modify the program to print the permutations in the opposite order, and also to output a permutation count on each line.
Run the program again and input the string cab. Check that the output is reversed.
Run the program again with an input string of abcdef. Why did the program take longer to produce the results?
OUTPUT:

Enter a string to permute (<Enter> to exit): 
cab
cba
acb
abc
bca
bac
Enter a string to permute (<Enter> to exit):