Create 4.7.2: Nested loop example: Histogram.c
Below is a nested loop example that graphically depicts an integer's magnitude by using asterisks, creating a "histogram." The inner loop is a for loop that handles the printing of the asterisks. The outer loop is a while loop that handles executing until a negative number is entered.
OUTPUT:


Enter an integer (negative to quit): 9
Depicted graphically:
*********

Enter an integer (negative to quit): 23
Depicted graphically:
***********************

Enter an integer (negative to quit): 25
Depicted graphically:
*************************

Enter an integer (negative to quit): -1
Goodbye.