Create 4.2.1: While loop.c
4.2 While loops
While loop: Basics
A while loop is a program construct that repeatedly executes a list of sub-statements (known as the loop body) while the loop's expression evaluates to true. Each execution of the loop body is called an iteration. Once entering the loop body, execution continues to the body's end, even if the expression would become false midway through.

Captions

When encountered, a while loop's expression is evaluated. If true, the loop's body is entered. Here, userChar was initialized with 'y', so userChar == 'y' is true.
Thus, the loop body is executed, which outputs currPower's current value of 2, doubles currPower, and gets the next input.
Execution jumps back to the while part. userChar is 'y' (the first input), so userChar == 'y' is true, and the loop body executes (again), outputting 4.
userChar is 'y' (the second user input), so userChar == 'y' is true, and the loop body executes (a third time), outputting 8.
userChar is now 'n', so userChar == 'y' is false. Thus, execution jumps to after the loop, which outputs "Done".